+++ lib/api_utils.dart
@@ -1,6 +1,7 @@
 library api_utils;
 
-export 'src/api_logger.dart';
+export 'src/api_config.dart';
 export 'src/api_response.dart';
 export 'src/request.dart';
 export 'src/status_code.dart';
+export 'src/timeout.dart';