+++ lib/src/request.dart
@@ -4,7 +4,8 @@ import 'dart:convert';
 // ignore: import_of_legacy_library_into_null_safe
 import 'package:http/http.dart' as http;
 import 'package:meta/meta.dart';
-import 'package:api_utils/src/api_logger.dart';
+import 'package:api_utils/src/api_config.dart';
+import 'package:api_utils/src/timeout.dart';
 import 'package:api_utils/src/status_code.dart';
 import 'package:api_utils/src/api_response.dart';
 
@@ -22,13 +23,16 @@ const Map<String, String> _defaultHeaders = {
 typedef FromJson<T> = T Function(Map<String, dynamic>);
 
 /// Make a GET request with a json object response
-Future<ApiResponse<T>> get<T>(
-    {required String url,
-    required FromJson<T> fromJson,
-    bool useFromJsonOnFailure = false,
-    Map<String, String> headers = _defaultHeaders}) async {
+Future<ApiResponse<T>> get<T>({
+  @required String url,
+  @required FromJson<T> fromJson,
+  bool useFromJsonOnFailure = false,
+  Map<String, String> headers,
+  Duration timeout,
+}) async {
   try {
-    var response = await _client.get(url, headers: headers);
+    var requestFuture = _client.get(url, headers: headers);
+    var response = await _makeRequest(requestFuture, timeout);
     return _handleResult('GET', url, response, fromJson, useFromJsonOnFailure);
   } on Exception catch (e, stack) {
     _onException('GET', url, -1, e, stack);
@@ -37,11 +41,14 @@ Future<ApiResponse<T>> get<T>(
 }
 
 /// Make a GET request with a byte array response
-Future<ApiResponse<Uint8List>> getByteArray(
-    {required String url,
-    Map<String, String> headers = _defaultHeaders}) async {
+Future<ApiResponse<Uint8List>> getByteArray({
+  @required String url,
+  Map<String, String> headers,
+  Duration timeout,
+}) async {
   try {
-    var response = await _client.get(url, headers: headers);
+    var requestFuture = _client.get(url, headers: headers);
+    var response = await _makeRequest(requestFuture, timeout);
     return _handleByteArrayResult('GET', url, response);
   } on Exception catch (e, stack) {
     _onException('GET', url, -1, e, stack);
@@ -50,12 +57,15 @@ Future<ApiResponse<Uint8List>> getByteArray(
 }
 
 /// Make a GET request with an json list response
-Future<ApiResponse<List<T>>> getList<T>(
-    {required String url,
-    required FromJson<T> fromJson,
-    Map<String, String> headers = _defaultHeaders}) async {
+Future<ApiResponse<List<T>>> getList<T>({
+  @required String url,
+  @required FromJson<T> fromJson,
+  Map<String, String> headers,
+  Duration timeout,
+}) async {
   try {
-    var response = await _client.get(url, headers: headers);
+    var requestFuture = _client.get(url, headers: headers);
+    var response = await _makeRequest(requestFuture, timeout);
     return _handleListResult('GET', url, response, fromJson);
   } on Exception catch (e, stack) {
     _onException('GET', url, -1, e, stack);
@@ -65,15 +75,18 @@ Future<ApiResponse<List<T>>> getList<T>(
 
 /// Make a POST request sending a json object, with an optional json object
 /// response
-Future<ApiResponse<T>> post<T>(
-    {required String url,
-    required Map<String, dynamic> body,
-    FromJson<T>? fromJson,
-    bool useFromJsonOnFailure = false,
-    Map<String, String> headers = _defaultHeaders}) async {
+Future<ApiResponse<T>> post<T>({
+  @required String url,
+  @required Map<String, dynamic> body,
+  FromJson<T> fromJson,
+  bool useFromJsonOnFailure = false,
+  Map<String, String> headers,
+  Duration timeout,
+}) async {
   try {
-    var response =
-        await _client.post(url, headers: headers, body: jsonEncode(body));
+    var requestFuture =
+        _client.post(url, headers: headers, body: jsonEncode(body));
+    var response = await _makeRequest(requestFuture, timeout);
     return _handleResult('POST', url, response, fromJson, useFromJsonOnFailure);
   } on Exception catch (e, stack) {
     _onException('POST', url, -1, e, stack);
@@ -83,14 +96,17 @@ Future<ApiResponse<T>> post<T>(
 
 /// Make a POST request sending a raw string, with an optional json object
 /// response
-Future<ApiResponse<T>> postAsString<T>(
-    {required String url,
-    required String body,
-    FromJson<T>? fromJson,
-    bool useFromJsonOnFailure = false,
-    Map<String, String> headers = _defaultHeaders}) async {
+Future<ApiResponse<T>> postAsString<T>({
+  @required String url,
+  @required String body,
+  FromJson<T> fromJson,
+  bool useFromJsonOnFailure = false,
+  Map<String, String> headers,
+  Duration timeout,
+}) async {
   try {
-    var response = await _client.post(url, headers: headers, body: body);
+    var requestFuture = _client.post(url, headers: headers, body: body);
+    var response = await _makeRequest(requestFuture, timeout);
     return _handleResult('POST', url, response, fromJson, useFromJsonOnFailure);
   } on Exception catch (e, stack) {
     _onException('POST', url, -1, e, stack);
@@ -100,14 +116,17 @@ Future<ApiResponse<T>> postAsString<T>(
 
 /// Make a POST request sending a json object, with an optional json list
 /// response
-Future<ApiResponse<List<T>>> postAndGetList<T>(
-    {required String url,
-    required Map<String, dynamic> body,
-    required FromJson<T> fromJson,
-    Map<String, String> headers = _defaultHeaders}) async {
+Future<ApiResponse<List<T>>> postAndGetList<T>({
+  @required String url,
+  @required Map<String, dynamic> body,
+  @required FromJson<T> fromJson,
+  Map<String, String> headers,
+  Duration timeout,
+}) async {
   try {
-    var response =
-        await _client.post(url, headers: headers, body: jsonEncode(body));
+    var requestFuture =
+        _client.post(url, headers: headers, body: jsonEncode(body));
+    var response = await _makeRequest(requestFuture, timeout);
     return _handleListResult('POST', url, response, fromJson);
   } on Exception catch (e, stack) {
     _onException('POST', url, -1, e, stack);
@@ -117,15 +136,18 @@ Future<ApiResponse<List<T>>> postAndGetList<T>(
 
 /// Make a PUT request sending a json object, with an optional json object
 /// response
-Future<ApiResponse<T>> put<T>(
-    {required String url,
-    required Map<String, dynamic> body,
-    FromJson<T>? fromJson,
-    bool useFromJsonOnFailure = false,
-    Map<String, String> headers = _defaultHeaders}) async {
+Future<ApiResponse<T>> put<T>({
+  @required String url,
+  @required Map<String, dynamic> body,
+  FromJson<T> fromJson,
+  bool useFromJsonOnFailure = false,
+  Map<String, String> headers,
+  Duration timeout,
+}) async {
   try {
-    var response =
-        await _client.put(url, headers: headers, body: jsonEncode(body));
+    var requestFuture =
+        _client.put(url, headers: headers, body: jsonEncode(body));
+    var response = await _makeRequest(requestFuture, timeout);
     return _handleResult('PUT', url, response, fromJson, useFromJsonOnFailure);
   } on Exception catch (e, stack) {
     _onException('PUT', url, -1, e, stack);
@@ -135,15 +157,18 @@ Future<ApiResponse<T>> put<T>(
 
 /// Make a PUT request sending a json list, with an optional json object
 /// response
-Future<ApiResponse<T>> putList<T>(
-    {required String url,
-    required List body,
-    FromJson<T>? fromJson,
-    bool useFromJsonOnFailure = false,
-    Map<String, String> headers = _defaultHeaders}) async {
+Future<ApiResponse<T>> putList<T>({
+  @required String url,
+  @required List body,
+  FromJson<T> fromJson,
+  bool useFromJsonOnFailure = false,
+  Map<String, String> headers,
+  Duration timeout,
+}) async {
   try {
-    var response =
-        await _client.put(url, headers: headers, body: jsonEncode(body));
+    var requestFuture =
+        _client.put(url, headers: headers, body: jsonEncode(body));
+    var response = await _makeRequest(requestFuture, timeout);
     return _handleResult('PUT', url, response, fromJson, useFromJsonOnFailure);
   } on Exception catch (e, stack) {
     _onException('PUT', url, -1, e, stack);
@@ -152,11 +177,14 @@ Future<ApiResponse<T>> putList<T>(
 }
 
 /// Make a DELETE request
-Future<ApiResponse<T>> delete<T>(
-    {required String url,
-    Map<String, String> headers = _defaultHeaders}) async {
+Future<ApiResponse<T>> delete<T>({
+  @required String url,
+  Map<String, String> headers,
+  Duration timeout,
+}) async {
   try {
-    var response = await _client.delete(url, headers: headers);
+    var requestFuture = _client.delete(url, headers: headers);
+    var response = await _makeRequest(requestFuture, timeout);
     return _handleResult('DELETE', url, response, null, false);
   } on Exception catch (e, stack) {
     _onException('DELETE', url, -1, e, stack);
@@ -166,15 +194,18 @@ Future<ApiResponse<T>> delete<T>(
 
 /// Make a PATCH request sending a json object, with an optional json object
 /// response
-Future<ApiResponse<T>> patch<T>(
-    {required String url,
-    required Map<String, dynamic> body,
-    FromJson<T>? fromJson,
-    bool useFromJsonOnFailure = false,
-    Map<String, String> headers = _defaultHeaders}) async {
+Future<ApiResponse<T>> patch<T>({
+  @required String url,
+  @required Map<String, dynamic> body,
+  FromJson<T> fromJson,
+  bool useFromJsonOnFailure = false,
+  Map<String, String> headers,
+  Duration timeout,
+}) async {
   try {
-    var response =
-        await _client.patch(url, headers: headers, body: jsonEncode(body));
+    var requestFuture =
+        _client.patch(url, headers: headers, body: jsonEncode(body));
+    var response = await _makeRequest(requestFuture, timeout);
     return _handleResult(
         'PATCH', url, response, fromJson, useFromJsonOnFailure);
   } on Exception catch (e, stack) {
@@ -236,6 +267,29 @@ ApiResponse<List<T>> _handleListResult<T>(
   }
 }
 
+// if a per request timeout or global timeout is set, timeout the future when
+// it is created.
+Future<http.Response> _makeRequest(
+  Future<http.Response> requestFuture,
+  Duration timeout,
+) async {
+  http.Response response;
+
+  // Prefer per request timeout over the global timeout
+  var _timeout = timeout ?? ApiUtilsConfig.timeout;
+  if (_timeout != null) {
+    response = await requestFuture.timeout(
+      _timeout,
+      onTimeout: () {
+        throw ApiTimeoutException();
+      },
+    );
+  } else {
+    response = await requestFuture;
+  }
+  return response;
+}
+
 void _onException(
   String method,
   String url,
@@ -244,7 +298,7 @@ void _onException(
   StackTrace? stack,
 ]) {
   var message = 'Api Error: $statusCode $method $url Error: ${e.toString()}';
-  ApiLogger.onErrorMiddleware.forEach((x) => {x(message, e, stack)});
+  ApiUtilsConfig.onErrorMiddleware?.forEach((x) => {x(message, e, stack)});
 }
 
 void _onError(
@@ -254,5 +308,5 @@ void _onError(
   String error,
 ) {
   var message = 'Api Error: $statusCode $method $url Error: $error';
-  ApiLogger.onErrorMiddleware.forEach((x) => {x(message, null, null)});
+  ApiUtilsConfig.onErrorMiddleware?.forEach((x) => {x(message, null, null)});
 }