+++ lib/src/timeout.dart
@@ -1,0 +1,4 @@
+class ApiTimeoutException implements Exception {
+  const ApiTimeoutException();
+  String toString() => "ApiTimeoutException";
+}